Controlador estandar, archivo para conexion a BD
