+++ foro-cucei-api/models/Users.php
@@ -1,61 +1,45 @@
 <?php
 
+  include_once('c:/xampp/htdocs/ForoCucei/foro-cucei-api/Database.php');
+
   class Users {
-
     function __construct(){
-      include('config.php');
-
-      $dsn = 'mysql:dbname='.DB_NAME.';host='.DB_HOST;
-
-      try{
-        $this->pdo = new PDO($dsn, DB_USER, DB_PASS);
-      }
-      catch (PDOException $e){
-        echo 'Connection failed: ' . $e->getMessage();
-      }
-
+      $this->connection = DataBase::connection();
     }
 
-    function ShowUsers(){
-
+    //function SignUp($data){
+      //return $this->db->insert('users', $data);
+    //  return $this->db->insert('users', $data);
+    //}
+    function Show(){
       $st = $this->pdo->prepare('SELECT * FROM users');
-
       $st->execute();
-
       $result = $st->fetchAll(PDO::FETCH_OBJ);
-
       return $result;
     }
-
-    function CreateUsers(){
-      $name = 'Damon';
-      $last = 'Hill';
-      $nick = 'Hilltop';
-      $mail = 'DamonH@mail.com';
+    function Create(){
+      $name = 'Tony';
+      $last = 'Stark';
+      $nick = 'Ironman';
+      $mail = 'Stark@mail.com';
       $pass = '+Carlos1';
       $passHash = password_hash($pass, PASSWORD_BCRYPT);
-
       $st = $this->pdo->prepare('INSERT INTO users(name, last, nick, mail, pass) VALUES (:name, :last, :nick, :mail, :pass)');
       $st->bindValue(":name", $name);
       $st->bindValue(":last", $last);
       $st->bindValue(":nick", $nick);
       $st->bindValue(":mail", $mail);
       $st->bindValue(":pass", $passHash);
-
       $st->execute();
-
       $result = $st->fetchAll(PDO::FETCH_OBJ);
-
       return $result;
     }
-
-    function UpdateUsers(){
+    function Update(){
       $id = 2;
       $name = 'Kimi';
       $last = 'Raikkonen';
       $nick = 'iceman';
       $mail = 'iceone@mail.com';
-
       $st = $this->pdo->prepare('UPDATE users
         SET `name` = :name,
             `last` = :last,
@@ -68,23 +52,16 @@
       $st->bindValue(":last", $last);
       $st->bindValue(":nick", $nick);
       $st->bindValue(":mail", $mail);
-
       $st->execute();
-
       $result = $st->fetchAll(PDO::FETCH_OBJ);
-
       return $result;
     }
-
-    function DeleteUsers(){
+    function Delete(){
       $id = 3;
       $st = $this->pdo->prepare('DELETE FROM users WHERE `iduser` = :id');
       $st->bindValue(":id", $id);
       $st->execute();
-
       $result = $st->fetchAll(PDO::FETCH_OBJ);
-
       return $result;
     }
-
   }
